+++ includes/admin/menu/tools/group-products.php
@@ -1,3 +1,0 @@
-<?php include_once ANAR_PLUGIN_PATH . 'includes/admin/menu/tools/tool-publish-drafts.php';?>
-
-<?php include_once ANAR_PLUGIN_PATH . 'includes/admin/menu/tools/tool-gallery.php';?>