+++ includes/admin/menu/tools/group-reset.php
@@ -1,1 +1,0 @@
-<?php include_once ANAR_PLUGIN_PATH . 'includes/admin/menu/tools/tool-reset.php';?>